Refactor: split response.go, fix naming, simplify encodeParams
Maintainability improvements:
- Extract checkResponse/stripeError into response.go (matches GitHub
  connector's file structure pattern)
- Fix inconsistent constant naming: livKeyPrefix → liveKeyPrefix
  (now consistent with rLiveKeyPrefix)
- Extract validKeyPrefixes into a package-level var to avoid
  rebuilding the slice on every hasValidPrefix() call
- Simplify encodeParams: url.Values.Encode() already sorts keys
  alphabetically, so the manual sort was redundant
- Verified no race conditions (go test -race passes)
